requirements.txt
aniso8601==10.0.1
Authlib==1.7.0
blinker==1.9.0
cachelib==0.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
cryptography==46.0.7
DateTime==6.0
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-OAuthlib==0.9.6
Flask-RESTful==0.3.10
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joserfc==1.6.4
MarkupSafe==3.0.3
oauthlib==2.1.0
pycparser==3.0
PyMySQL==1.1.2
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
requests-oauthlib==1.1.0
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.7
zope.interface==8.2
